Move ws tests to their own file
